// server/researchindicators/src/domain/entities/impersonation/enum/impersonation-error-code.enum.ts
export enum ImpersonationErrorCodeEnum {
  NOT_ALLOWED = 'NOT_ALLOWED',
  NESTED = 'NESTED',
  SESSION_INVALID = 'SESSION_INVALID',
  SESSION_HEADER_REQUIRED = 'SESSION_HEADER_REQUIRED',
  TARGET_NOT_FOUND = 'TARGET_NOT_FOUND',
  TARGET_IS_ADMIN = 'TARGET_IS_ADMIN',
  TARGET_IS_SELF = 'TARGET_IS_SELF',
}